requirements.txt
timm==0.4.5
wget==3.2
h5py==3.6.0
librosa==0.8.1
matplotlib==3.5.1
museval==0.4.0
numpy==1.22.0
pandas==1.4.0
scikit_learn==1.0.2
scipy==1.7.3
tensorboard==2.8.0
torchcontrib==0.0.2
torchlibrosa==0.0.9
tqdm==4.62.3
ml-collections
nnAudio
einops
easydict
# gdown